Import Service/requirements.txt
aws-cdk-lib==2.99.1
constructs>=10.0.0
pytest~=7.4.0
pytest-mock~=3.11.1
boto3~=1.28.0
